Render data in modal + final touch ups
