+++ README.md
@@ -6,15 +6,14 @@ React, Redux, Express and Node app - converts a given numeric string into a list
 
 ## Demo
 
-...
+[Demo](/Users/JaiJk/Projects/number-to-word-list-convertor/translat9r-demo.mov)
 
 ### Why and How
 
 My goals for this project were to practise and demonstrate my abilities as a Full Stack Node.js & React Engineer. I did this by attempting to do the following:
 
 1. Writing high quality, readable, and scalable business logic, algorithms and backend REST APIs (Node.js, Express.js).
-2. Hooking those APIs up to a pleasing UI (React)
-3. Storing and Accessing State using Redux.
+2. Hooking those APIs up to a pleasing UI (React, Material UI, Axios)
 
 I also tried to mimic the process of a Commercial team as closely as possible whilst adhering to the nature of a solo project - I created, reviewed and merged Branches and Pull Requests.
 
@@ -22,9 +21,7 @@ In a Production environment, I would rebase my commits before pushing a PR or me
 
 ## How to Use
 
-Access Via Heroku here: ...
-
-To access locally:
+To run locally:
 
 1. Git clone: <br>
    `git@github.com:jai-jk/number-to-word-list-convertor.git` (SSH) <br>
@@ -41,23 +38,31 @@ To access locally:
 - Words are matched according to a collection of the 500 most common words in the English Language, and any words generated that match to those are displayed. All of the other possible words are also displayed.
 - Users can only enter numbers and the textfield is required - a flash notice is triggered if a user tries to circumvent either process.
 
+## Improvements I would have loved to make if I'd had more time
+
+- A More accessible and generally improved UI and UX - scalability for multiple screen sizes, ease of use for all types of users, improvements to the modal-close/refresh functionality, definitely refactoring of css and styles into external sheets wherever possible etc.
+- A functioning keyboard on the brick phone - what a missed opportunity, argh!!
+- Deployment to Heroku - just couldn't get it working in time =(.
+- Unit, Functional, Feature and E2E Tests with Jest and Cypress (TDD from the start would have been even better).
+- Migration to TypeScript.
+
 ### Dependencies
 
 Server-Side Dependencies:
 
-- Cors
-- Express
-- Morgan
+- [Cors](https://www.npmjs.com/package/cors)
+- [Express](https://www.npmjs.com/package/express)
+- [Morgan](https://www.npmjs.com/package/morgan)
 
 Client-Side Dependencies:
 
-- emotion/react
-- emotion/styled
-- material-ui/core
-- mui/icons-material
-- mui/material
-- axios
-- react
-- react-dom
-- react-scripts
-- styled-components
+- [emotion/react](https://www.npmjs.com/package/@emotion/react)
+- [emotion/styled](https://www.npmjs.com/package/@emotion/styled)
+- [material-ui/core](https://www.npmjs.com/package/@material-ui/core)
+- [mui/icons-material](https://www.npmjs.com/package/@mui/icons-material)
+- [mui/material](https://www.npmjs.com/package/@mui/material)
+- [axios](https://www.npmjs.com/package/axios)
+- [react](https://www.npmjs.com/package/react)
+- [react-dom](https://www.npmjs.com/package/react-dom)
+- [react-scripts](https://www.npmjs.com/package/react-scripts)
+- [styled-components](https://www.npmjs.com/package/styled-components)
