+++ client/src/components/Header.js
@@ -1,4 +1,5 @@
 import React from 'react';
+
 import icon from '../icon.svg';
 
 export const Header = () => {
